Implements product details retrieval
Adds the necessary services, DTOs, queries and repositories
to retrieve product details with brand and category names.

This change introduces the `ProductDetailDto` for data transfer,
an `IProductService` interface and its implementation `ProductServices`,
a `DataBaseException` class, and updates the product repository
to retrieve product details with related brand and category
information using a new query. It also refactors the web layer
to show brand and category names in product listing.

Updates the ProductViewModel to a record.

Adds bootstrap to the product index view for better UX.